MenuV2
The Menu flow for the Chat and Call button sequences are fixed. Even
though I haven't tried the userArray mechanism for both sequences, all
the buttons are rightfuly connected and the GUI flows gratefully. I
will verify the sequences with dummy lists and different lengths and
comment this menu code later. I have tested it on Phonegap on my
Android, and it works perfectly!