2 Days worth of extreme changes- iOS and Android Chat clients
I implemented chat client for the iOS with the client side code in the
application itself, mostly because iOS does not seem to have any
problems with any of the communication protocols used in SockJS. Now,
in Android's case, after the Adnroid Webviews failing every single
communication attempt by SockJS, I had to implement an inAppBrowser,
mostly because the broswer has no problems with the chat communication.
The good thing is, that code can be injected into this browser, making
value-passing possible in future chat sessions. The reason behind the
lack of commits was the various testing and modifications within SockJS
to make it work with Android, with no avail. Thankfully, the
inAppBroswer works for the chat.